requirements.txt
greenlet==2.0.2
SQLAlchemy==2.0.15
typing_extensions==4.6.3

aiogram~=2.25.1
cachetools~=5.3.1
python-dotenv~=1.0.0